MLProj/requirements.txt
# pip install -r requirements.txt


hdbscan
numpy
opencv-python
pandas
Pillow
scikit-learn
torch
torchvision
tqdm
umap-learn
transformers
